requirements.txt
ultralytics>=8.0.0
pyyaml
opencv-python
torch
Pillow
fastapi
uvicorn[standard]
python-multipart
